Top left tagline and image position fix.
Added some CSS and xs-hidden to the logo in the top left to fix position on tablet sized devices and work on wide screen Chrome browser. The positioning in Chrome was breaking and not displaying as absolute. The changed lines have been marked in the CSS file for testing purposes.